+++ src/SAWScript/Builtins.hs
@@ -1062,6 +1062,9 @@ proveMathSAT = proveSBV SBV.mathSAT
 proveYices :: ProofScript ()
 proveYices = proveSBV SBV.yices
 
+proveUnintBitwuzla :: [String] -> ProofScript ()
+proveUnintBitwuzla = proveUnintSBV SBV.bitwuzla
+
 proveUnintBoolector :: [String] -> ProofScript ()
 proveUnintBoolector = proveUnintSBV SBV.boolector
 
@@ -1085,6 +1088,9 @@ proveUnintYices = proveUnintSBV SBV.yices
 w4_abc_smtlib2 :: ProofScript ()
 w4_abc_smtlib2 = wrapW4Prover ABC [W4_SMTLib2] Prover.proveWhat4_abc []
 
+w4_bitwuzla :: ProofScript ()
+w4_bitwuzla = wrapW4Prover Bitwuzla [] Prover.proveWhat4_bitwuzla []
+
 w4_boolector :: ProofScript ()
 w4_boolector = wrapW4Prover Boolector [] Prover.proveWhat4_boolector []
 
@@ -1099,6 +1105,9 @@ w4_cvc5 = wrapW4Prover CVC5 [] Prover.proveWhat4_cvc5 []
 
 w4_yices :: ProofScript ()
 w4_yices = wrapW4Prover Yices [] Prover.proveWhat4_yices []
+
+w4_unint_bitwuzla :: [String] -> ProofScript ()
+w4_unint_bitwuzla = wrapW4Prover Bitwuzla [] Prover.proveWhat4_bitwuzla
 
 w4_unint_boolector :: [String] -> ProofScript ()
 w4_unint_boolector = wrapW4Prover Boolector [] Prover.proveWhat4_boolector