+++ src/SAWScript/Prover/What4.hs
@@ -109,7 +109,8 @@ proveExportWhat4_sym solver hashConsing outFilePath satq =
      -- Assume unsat
      return (Nothing, solver_name)
 
-proveWhat4_z3, proveWhat4_boolector,
+proveWhat4_z3,
+  proveWhat4_bitwuzla, proveWhat4_boolector,
   proveWhat4_cvc4, proveWhat4_cvc5,
   proveWhat4_dreal, proveWhat4_stp, proveWhat4_yices,
   proveWhat4_abc ::
@@ -118,6 +119,7 @@ proveWhat4_z3, proveWhat4_boolector,
   TopLevel (Maybe CEX, String)
 
 proveWhat4_z3        = proveWhat4_sym z3Adapter
+proveWhat4_bitwuzla  = proveWhat4_sym bitwuzlaAdapter
 proveWhat4_boolector = proveWhat4_sym boolectorAdapter
 proveWhat4_cvc4      = proveWhat4_sym cvc4Adapter
 proveWhat4_cvc5      = proveWhat4_sym cvc5Adapter
@@ -139,7 +141,8 @@ proveWhat4_z3_using tactic hashConsing satq =
           _ <- setOpt z3TacticSetting $ Text.pack tactic
           return ()
 
-proveExportWhat4_z3, proveExportWhat4_boolector,
+proveExportWhat4_z3,
+  proveExportWhat4_bitwuzla, proveExportWhat4_boolector,
   proveExportWhat4_cvc4, proveExportWhat4_cvc5,
   proveExportWhat4_dreal, proveExportWhat4_stp, proveExportWhat4_yices ::
   Bool          {- ^ Hash-consing of ExportWhat4 terms -}->
@@ -148,6 +151,7 @@ proveExportWhat4_z3, proveExportWhat4_boolector,
   TopLevel (Maybe CEX, String)
 
 proveExportWhat4_z3        = proveExportWhat4_sym z3Adapter
+proveExportWhat4_bitwuzla  = proveExportWhat4_sym bitwuzlaAdapter
 proveExportWhat4_boolector = proveExportWhat4_sym boolectorAdapter
 proveExportWhat4_cvc4      = proveExportWhat4_sym cvc4Adapter
 proveExportWhat4_cvc5      = proveExportWhat4_sym cvc5Adapter